Tidying code
